==> 06.RestAPI/01_roadapi.ipynb <==
# 행정안정부 도로명주소 API

## 건물명으로부터 도로명주소 구하기

import pandas as pd
import numpy as np

import requests
from urllib.parse import urlparse, quote

key_fd = open('roadapikey.txt', mode='r')
roadapi_key = key_fd.read(100)
key_fd.close()

road_url = 'https://www.juso.go.kr/addrlink/addrLinkApi.do'
confmKey = roadapi_key
currentPage = 1
countPerPage = 5
keyword = quote('강서구청')
resultType = 'json'

request_url = f"{road_url}?confmKey={roadapi_key}&currentPage={currentPage}&countPerPage={countPerPage}&keyword={keyword}&resultType=json"

urlparse(request_url).geturl()

quote('강서구청')

request_url = f"{road_url}?confmKey={roadapi_key}&currentPage={currentPage}&countPerPage={countPerPage}&keyword={keyword}&resultType=json"

results = requests.get(urlparse(request_url).geturl()).json()
addr_list = results['results']['juso']
addr_list[0]

for addr in addr_list:
    print(addr['bdNm'], addr['roadAddr'])

## DataFrame

def getAPI(**kwargs):
    confmKey = roadapi_key
    baseUrl = "https://www.juso.go.kr/addrlink/addrLinkApi.do?confmKey="
    params = []

    for k, v in kwargs.items():
        params.append("{}={}".format(k, v))        

    param = '&' + '&'.join(params)
    query = baseUrl + confmKey + param
    
    return requests.get(urlparse(query).geturl()).json()

data1 = getAPI(currentPage=1, countPerPage=5, keyword='강서구청', resultType='json')
data2 = getAPI(currentPage=1, countPerPage=5, keyword='양천구청', resultType='json')
data3 = getAPI(currentPage=1, countPerPage=5, keyword='부천시청', resultType='json')
data4 = getAPI(currentPage=1, countPerPage=5, keyword='영등포구청', resultType='json')
data5 = getAPI(currentPage=1, countPerPage=5, keyword='김포시청', resultType='json')

juso = [
    data1['results']['juso'][0],
    data2['results']['juso'][0],
    data3['results']['juso'][0],
    data4['results']['juso'][0],
    data5['results']['juso'][0]
]
df = pd.DataFrame(juso)
df

df.to_csv('address.csv')
