==> 06.RestAPI/02_kakaoapi.ipynb <==
# Kakao Map API

이 문서는 로컬(Local) API 구현 방법을 소개합니다.   

<https://developers.kakao.com/docs/latest/ko/local/dev-guide>

import pandas as pd
import numpy as np

import requests
from urllib.parse import urlparse

key_fd = open('kakaomaprestapikey.txt', mode='r')
API_KEY = key_fd.read(100)
key_fd.close()

def getAPI(**kwargs):
    baseUrl = "https://dapi.kakao.com/v2/local/search/address.json?"
    params = []

    for k, v in kwargs.items():
        params.append("{}={}".format(k, v))        

    param = '&'.join(params)
    query = baseUrl + param

    return requests.get(urlparse(query).geturl(), headers={"Authorization": "KakaoAK "+API_KEY}).json()

data1 = getAPI(query='서울특별시 강서구 화곡로 302 (화곡동)')
data1['documents'][0]

lng = float(data1['documents'][0]['x'])
lat = float(data1['documents'][0]['y'])
lng, lat

df = pd.read_csv('address.csv')
df

def getCoords(address):
    api = getAPI(query=address)
    coords = {}
    coords.update({ 'x': float(api['documents'][0]['x']) })
    coords.update({ 'y': float(api['documents'][0]['y']) })
    return coords

print(getCoords('서울특별시 강서구 화곡로 302 (화곡동)'))

df[:1]

df['lat'] = np.nan
df['lon'] = np.nan

for i in df.index:
    addr = df['roadAddr'][i]
    coords = getCoords(addr)
    df['lat'][i] = coords['y']
    df['lon'][i] = coords['x']

df

df.to_csv('address2.csv', sep=',', encoding='utf8', index=False)
