==> 06.RestAPI/03_openweather.ipynb <==
# Open Weather API

## Current Weather By geographic coordinates

<https://openweathermap.org/current>

import pandas as pd
import numpy as np

import requests
from urllib.parse import urlparse

key_fd = open('openweatherapikey.txt', mode='r')
API_KEY = key_fd.read(100)
key_fd.close()

lat = 37.550966
lon = 126.849532
units = 'metric'
lang = 'kr'

query = f"http://api.openweathermap.org/data/2.5/weather?lat={lat}&lon={lon}&units={units}&lang={lang}&appid={API_KEY}"

result = requests.get(urlparse(query).geturl()).json()

main = result['weather'][0]['main']
desc = result['weather'][0]['description']
main, desc

temp = result['main']['temp']
temp_min = result['main']['temp_min']
temp_max = result['main']['temp_max']
temp, temp_min, temp_max

temp = round(float(temp)+0.01, 1)
temp

### address2.csv 파일 읽어서 날씨 정보 입력하기

df = pd.read_csv('address2.csv')
df

weather_list = []; desc_list = []; temp_list = []

for i in df.index:
    url = f"http://api.openweathermap.org/data/2.5/weather?lat={lat}&lon={lon}&appid={API_KEY}&units=metric&lang=kr"

    result = requests.get(urlparse(url).geturl()).json()
    main = result['weather'][0]['main']
    desc = result['weather'][0]['description']
    temp = result['main']['temp']
    temp = round(float(temp)+0.01, 1)
    temp_min = result['main']['temp_min']
    temp_max = result['main']['temp_max']
    weather_list.append(main)
    desc_list.append(desc)
    temp_list.append(f"{temp}/{temp_min}/{temp_max}")

df['weather'] = weather_list
df['desc'] = desc_list
df['temp'] = temp_list
df

df.to_csv('address3.csv', sep=',', encoding='utf8', index=False)
