==> 06.RestAPI/04_folium.ipynb <==
# Folium 지도를 이용하여 표시하기

### Folium 한글 깨짐 현상 해결하기

**branca 라이브러리 수정 사항 반영**   
```
$ pip install git+https://github.com/python-visualization/branca.git@master
```

**HTML 파일로 저장 후 확인하기**   
```
$ m.save("visualization.html")
```

### Reference
<https://somjang.tistory.com/entry/Python-Folium-한글-깨짐-현상-해결하기>   

import numpy as np
import pandas as pd
import folium

folium.__version__

mapping = folium.Map(location=[37.550966, 126.849532], zoom_start=13)
mapping

### Tiles

mapping = folium.Map(location=[37.550966, 126.849532], zoom_start=13, tiles='Stamen Toner')
mapping

### Marker

mapping = folium.Map(location=[37.550966, 126.849532], zoom_start=12)

folium.Marker(
    location=[37.517075, 126.866542],
    popup='양천구청',
    icon=folium.Icon(icon='cloud')
).add_to(mapping)

folium.Marker(
    location=[37.550966, 126.849532],
    popup='강서구청',
    icon=folium.Icon(color='green')
).add_to(mapping)

folium.Marker(
    location=[37.502277, 126.765003],
    popup='구천시청',
    icon=folium.Icon(color='red', icon='info-sign')
).add_to(mapping)

mapping

### address3.csv를 읽어서 지도 위에 표시하기

df = pd.read_csv('address3.csv')
df

df['icon'] = ['map-marker', 'cloud', 'gift', 'info-sign', 'ok-circle']
df

# 데이터의 중심좌표 구하기
mapping = folium.Map(location=[df.lat.mean(), df.lon.mean()], zoom_start=13)
mapping

# 날씨에 따른 마커의 색깔
color_dict = {
    'clear': 'blue',
    'Snow': 'white',
    'Rain': 'gray',
    'Extreme': 'red',
    'Clouds': 'orange',
    'Mist': 'green'
}

for i in df.index:
    print(df.lat[i], df.lon[i])
    folium.Marker(
        location=[df.lat[i], df.lon[i]],
        popup=df.bdNm[i],
        tooltip=f"{df.desc[i]}, {df.temp[i]}",
        icon=folium.Icon(color=color_dict[df.weather[i]], icon=df.icon[i])
    ).add_to(mapping)

mapping
