==> 06.RestAPI/05_flask.ipynb <==
# Flask

**Installation**   
```
(base) conda install -c anaconda flask
```

**Reference**   
<https://anaconda.org/anaconda/flask>

## Using folium with flask

<https://python-visualization.github.io/folium/flask.html>

## flask_example.py

""" flask_example.py

    Required packages:
    - flask
    - folium

    Usage:

    Start the flask server by running:

        $ python flask_example.py

    And then head to http://127.0.0.1:5000/ in your browser to see the map displayed

"""

from flask import Flask

import folium

app = Flask(__name__)


@app.route('/')
def index():
    start_coords = (46.9540700, 142.7360300)
    folium_map = folium.Map(location=start_coords, zoom_start=14)
    return folium_map._repr_html_()


if __name__ == '__main__':
    app.run(debug=True)
